+++ edgy/core/db/fields/core.py
@@ -5,7 +5,7 @@ import ipaddress
 import uuid
 from collections.abc import Callable, Sequence
 from enum import EnumMeta
-from functools import partial
+from functools import cached_property, partial
 from re import Pattern
 from typing import TYPE_CHECKING, Any, Optional, Union, cast
 
@@ -16,7 +16,7 @@ from pydantic import EmailStr
 from edgy.core.db.context_vars import CURRENT_INSTANCE, CURRENT_PHASE, EXPLICIT_SPECIFIED_VALUES
 from edgy.core.db.fields._internal import IPAddress
 from edgy.core.db.fields._validators import IPV4_REGEX, IPV6_REGEX
-from edgy.core.db.fields.base import Field
+from edgy.core.db.fields.base import BaseField, Field
 from edgy.core.db.fields.factories import FieldFactory
 from edgy.core.db.fields.types import BaseFieldType
 from edgy.exceptions import FieldDefinitionError
@@ -28,6 +28,65 @@ if TYPE_CHECKING:
 
 
 CLASS_DEFAULTS = ["cls", "__class__", "kwargs"]
+
+
+class ComputedField(BaseField):
+    def __init__(
+        self,
+        getter: Union[Callable[["BaseModelType", str], Any], str],
+        setter: Union[Callable[["BaseModelType", str, Any], None], str, None] = None,
+        fallback_getter: Optional[Callable[["BaseModelType", str], Any]] = None,
+        **kwargs: Any,
+    ) -> None:
+        kwargs["default"] = None
+        kwargs["field_type"] = kwargs["annotation"] = Any
+        self.getter = getter
+        self.fallback_getter = fallback_getter
+        self.setter = setter
+        super().__init__(
+            **kwargs,
+        )
+
+    @cached_property
+    def compute_getter(self) -> Callable[["BaseModelType", str], Any]:
+        if isinstance(self.getter, str):
+            fn = cast(
+                Optional[Callable[["BaseModelType", str], Any]],
+                getattr(self.owner, self.getter, None),
+            )
+        else:
+            fn = self.getter
+        if fn is None and self.fallback_getter is not None:
+            fn = self.fallback_getter
+        if fn is None:
+            raise ValueError(f"No getter found for attribute: {self.getter}.")
+        return fn
+
+    @cached_property
+    def compute_setter(self) -> Callable[["BaseModelType", str, Any], None]:
+        if isinstance(self.setter, str):
+            fn = cast(
+                Optional[Callable[["BaseModelType", str, Any], None]],
+                getattr(self.owner, self.setter, None),
+            )
+        else:
+            fn = self.setter
+        if fn is None:
+            return lambda instance, name, value: None
+        return fn
+
+    def to_model(
+        self,
+        field_name: str,
+        value: Any,
+    ) -> dict[str, Any]:
+        return {}
+
+    def __get__(self, instance: "BaseModelType", owner: Any = None) -> Any:
+        return self.compute_getter(instance, self.name)
+
+    def __set__(self, instance: "BaseModelType", value: Any) -> None:
+        self.compute_setter(instance, self.name, value)
 
 
 class CharField(FieldFactory, str):